# frontend/types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Project Desert API",
    "description": "API for managing ascetical practices and spiritual growth",
    "version": "2.0.0"
  },
  "paths": {
...
    "/admin/users": {
      "get": {
        "tags": ["admin"],
        "summary": "Get All Users",
        "description": "Get all users with their details and activity counts.\nRequires admin authentication.",
        "operationId": "get_all_users_admin_users_get",
        "parameters": [
          {
            "name": "x-user-email",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "X-User-Email"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/UserResponse" },
                  "title": "Response Get All Users Admin Users Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/admin/users/role": {
      "post": {
        "tags": ["admin"],
        "summary": "Update User Role",
        "description": "Update a user's role.\nPrevents admins from demoting themselves.",
        "operationId": "update_user_role_admin_users_role_post",
        "parameters": [
          {
            "name": "x-user-email",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "X-User-Email"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateRoleRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/admin/users/ban": {
      "post": {
        "tags": ["admin"],
        "summary": "Toggle User Ban",
        "description": "Ban or unban a user.\nPrevents admins from banning themselves.",
        "operationId": "toggle_user_ban_admin_users_ban_post",
        "parameters": [
          {
            "name": "x-user-email",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "X-User-Email"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ToggleBanRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/admin/current-user": {
      "get": {
        "tags": ["admin"],
        "summary": "Get Current User",
        "description": "Get current user info including role and ban status.",
        "operationId": "get_current_user_admin_current_user_get",
        "parameters": [
          {
            "name": "x-user-email",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "X-User-Email"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CurrentUserResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CurrentUserResponse": {
        "properties": {
          "id": { "type": "integer", "title": "Id" },
          "name": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Name"
          },
          "email": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Email"
          },
          "role": { "type": "string", "title": "Role" },
          "isBanned": { "type": "boolean", "title": "Isbanned" }
        },
        "type": "object",
        "required": ["id", "name", "email", "role", "isBanned"],
        "title": "CurrentUserResponse",
        "description": "Current user information."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ToggleBanRequest": {
        "properties": {
          "userId": { "type": "integer", "title": "Userid" },
          "isBanned": { "type": "boolean", "title": "Isbanned" }
        },
        "type": "object",
        "required": ["userId", "isBanned"],
        "title": "ToggleBanRequest",
        "description": "Request to ban or unban a user."
      },
...
      "UpdateRoleRequest": {
        "properties": {
          "userId": { "type": "integer", "title": "Userid" },
          "newRole": { "type": "string", "title": "Newrole" }
        },
        "type": "object",
        "required": ["userId", "newRole"],
        "title": "UpdateRoleRequest",
        "description": "Request to update user role."
      },
...
      "UserResponse": {
        "properties": {
          "id": { "type": "integer", "title": "Id" },
          "name": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Name"
          },
          "email": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Email"
          },
          "image": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Image"
          },
          "role": { "type": "string", "title": "Role" },
          "isBanned": { "type": "boolean", "title": "Isbanned" },
          "emailVerified": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Emailverified"
          },
          "userAsceticismsCount": {
            "type": "integer",
            "title": "Userasceticismscount"
          },
          "groupMembersCount": {
            "type": "integer",
            "title": "Groupmemberscount"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "email",
          "image",
          "role",
          "isBanned",
          "emailVerified",
          "userAsceticismsCount",
          "groupMembersCount"
        ],
        "title": "UserResponse",
        "description": "User response with activity counts."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}